requirements.txt
pulumi
pulumi-aws
python-dotenv
pulumi-automation
pulumi-std
